fortune_bot/requirements.txt
google-genai
moviepy
google-api-python-client
google-auth-oauthlib
python-dotenv
apscheduler
requests
